Moved some button components
